[DTensor] Move op_strategy_context to common_dtensor module
Move `op_strategy_context` from `test_op_strategy.py` to the shared `common_dtensor.py` testing utilities module, reducing code duplication across test files.

Fix the import error in D97796338.

Differential Revision: [D98008406](https://our.internmc.facebook.com/intern/diff/D98008406/)

[ghstack-poisoned]
